fix int overflow in bench. add benchmark script to store old benchmark results

+++ src/uci_interface.cpp
@@ -170,12 +170,12 @@ void UCIInterface::process_bench_command(std::string command) {
     int threads = std::stoi(parts[depthloc + 1]);
     UCIInterface::uci_response("Generating moves to depth: " + std::to_string(depth));
     auto start = std::chrono::high_resolution_clock::now();
-    int nummoves = movegen_benchmark::gen_num_moves(Game::instance().get_state(), depth, -1);
+    int64_t nummoves = movegen_benchmark::gen_num_moves(Game::instance().get_state(), depth, -1);
     auto stop = std::chrono::high_resolution_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(stop - start);
     UCIInterface::uci_response(std::to_string(nummoves) + " nodes found at this depth.");
     UCIInterface::uci_response("Time taken: " + std::to_string(duration.count()) + " ms.");
-    int mps = (1000 * nummoves / duration.count());
+    int64_t mps = (1000 * nummoves / duration.count());
     UCIInterface::uci_response("Nodes per second: " + std::to_string(mps));
 }
